Move `aiida.cmdline.utils.common.get_database_summary` to storage backend
This utility was used in `verdi archive inspect` and `verdi storage info`
to give information about the contents of an archive or normal storage
backend. Historically, a distinction used to be made between the contents
of the database and the repository, but in v2.0 these are unified by the
storage backend. The information about the repository is already
retrieved through `StorageBackend.get_info` but the contents of the
database were still built through the `get_database_summary` function.

The implementation is moved to the `get_orm_entity_overview` method of
the `StorageBackend` base class. This is done because the implementation
is currently still independent of the storage backend implementaiton. So
for now the implementations can simply call to this method in their
`get_info` implementation, but it leaves the option to override it with
a more performant implementation that doesn't go through the shared ORM
of AiiDA.