Profile: formalize what constitutes a test profile
Up till now, a profile was defined as a test profile if its name started
with `test_`. However, this is neither very robust nor very flexible as
it restricts naming for profiles.

Here we make it more explicit by adding the key `test_profile` to all
profiles through a configuration migration. Profiles whose name starts
with `test_` have `test_profile` set to `True`.

Profiles that have `test_profile = True` are considered a test profile
and so are allowed to be used in running the test suite. For safety
reasons, the `Profile.is_test_profile` property returns `True` if and
only if `test_profile` is set to `True`. If it is missing or any other
value is set, it returns `False`. This prevents a profile accidentally
being used for tests.