fix: better error message when flask not installed
The dependencies for the AiiDA REST API are currently not installed by
default - only when users request the `rest` extra.

When users were starting the REST API without installing the extra, they
were greeted by an unhelpful/unexpected error message

	ModuleNotFoundError: No module named 'flask_restful'

Here we catch ImportErrors upon starting the REST API and provide an
error message that includes instructions on how to remedy the situation.
For cases where the ImportError has another cause, the original
exception is carried along as well.