`ProfileManager`: always clear storage in `clear_profile`
The call to `_clear` on the storage in `ProfileManager.clear_profile`
was conditional on it already having been loaded. However, sometimes the
method may be called before the storage had already been loaded. This is
the case, for example, for the `aiida_profile_clean` fixture, which is
used by tests to start with an empty storage. Due to the conditional, in
some cases the storage would not be emptied and the test would fail if
the storage still contained data from a previous session.