Correct version number for RabbitMQ warning (#5413)
Recently, a warning was added when a RabbitMQ installation was used with
a version of 3.8 or higher. This was done because a consumer default
timeout was added to the server configuration that would cause unacked
messages to be requeued after only 15 minutes. However, this change was
only added in 3.9:

    https://github.com/rabbitmq/rabbitmq-server/commit/80e399212eb35d24969e869d80db60b706e59af3

and was backported in the v3.8 minor version in v3.8.15. Therefore, we
can relax the condition for the warning to be shown only when the
version is 3.8.15 or higher.