`GroupParamType`: store group if created (#5411)
When `create_if_not_exist` is equal to `True`, the `GroupParamType` will
create the group if it is not found when `convert` is called. The
returned group however was not stored forcing the caller to store it
themselves. When a client sets `create_if_not_exist` for a CLI command
it will be extremely unlikely that they don't want the group to be
stored.