`Config`: add the `warnings.rabbitmq_version` option (#5415)
This option can be used to suppress the warning that is emitted when a
profile storage is loaded that is configured with a version of RabbitMQ
that is known to be able to cause problems.

The `check_rabbitmq_version` and `check_version` free functions are
moved to `Manager`.

The `is_rabbitmq_version_supported` function is removed as it is no
longer used.

Cherry-pick: 9dd216ff510ac0a30a4b2020ac0875b266d52b28