`Config`: add the hidden `warnings.rabbitmq_version` option (#5415)
This option can be used to suppress the warning that is emitted when a
profile storage is loaded that is configured with a version of RabbitMQ
that is known to be able to cause problems. The option is hidden because
the warning is very important and should not easily be ignorable by
normal users, but it is mostly intended for developers.

The `check_rabbitmq_version` and `check_version` free functions are
moved to `Manager` as methods which allows to use `get_option` instead
of the `get_config_option` free function which is just an indirect way
of calling `Manager.get_option`.

The `is_rabbitmq_version_supported` function is also corrected. It was
still using `< parse('3.8')` as a condition whereas it should be
`< parse('3.8.15')` which was recently already changed for the
`check_rabbitmq_version` function.