feat: make aiida ipython magics available
AiiDA comes with an `%aiida` IPython magic that loads the default
profile (and imports some modules).
However, in order to take advantage of this, so far, users had to
manually copy a python module into their IPython startup folder.

This change makes it possible to load the ipython magics as follows:

%load_ext aiida
%aiida