CLI: fix shell completion of script for `verdi run`
When we moved to the built-in shell completion of `click==8.0` instead
of the external `click_completion`, the script argument of `verdi run`
was no longer auto-completed. The reason is that `click` does not
autocomplete files for the `STRING` type.

The parameter type is changed to `click.Path`, which not only fixes the
problem of auto-completion, it also provides automatic validation by
making sure if it is an existing file and not a directory. This allows
to simplify the logic in the command body significantly as well as we no
longer need to handle the `IOError` of unreadable files manually.