CLI: monkeypatch `click` to fix shell-completion for relative filepaths
Since `click==8.0` started shipping with their own implementation of
shell completion, we started using that instead of `click_completion`.
However, `click`'s implementation currently does not auto-complete
relative filepaths that start, for example, with either `.`, `~` or `$`.

They made an exception for the `/` character, in the following PR:

    https://github.com/pallets/click/pull/1930

Other users have noticed the problem with relative filepaths:

    https://github.com/pallets/click/issues/2040

But it is not sure if they plan to support this anytime soon so we
simply monkeypatch `click` ourselves by changing the implementation for
the `click.shell_completion._start_of_option`.