Docs: improve the backup section for v2.0
With AiiDA v2.0 the repository is now provided by the `disk-objectstore`
which is optimized for being backed up with `rsync`. This means the old
custom implementation of backing it up is no longer necessary nor
supported. It can now simply be backed up using `rsync`.

The instructions are updated and improved by tying the backup procedure
to the storage backend used for the profile. This will allow to have
other instructions for other storage backends in the future, as they
will most likely not be the same.