`load_backend_if_not_loaded`: test `get_profile_storage` is called just once
Add a test for `aiida.cmdline.utils.decorators.load_backend_if_not_loaded`
to make sure that it calls `Manager.get_profile_storage` only when a
profile and its storage haven't already been loaded.

The most naive solution would be to simply load a profile and storage,
mock `Manager.get_profile_storage` to raise, and then check that calling
the decorator raises. However, this runs the risk that if the decorator
is ever change to not call `Manager.get_profile_storage` the test would
pass, but not for the right reasons. We have to make sure before testing
`get_profile_storage` is not calle twice, that it is called once to
begin with.

This is accomplished by mocking the method first and checking that it
actually gets called, before loading the profile storage directly and
then calling the decorator again, whilst checking the mocked method was
not called again.