CLI: fix prompt info being shown in `verdi code setup --config` (#5434)
When calling `verdi code setup` with the `--config` option providing a
config file that specifies all required options, the generic prompting
info would still be displayed. Technically this is not a bug since the
command is being invoked in interactive mode and so it is not wholly
surprising that the generic prompt info is displayed.

The prompt actually comes from the `--code-rel-path` and `--code-folder`
options (at least when configuring a remote code) which are in that case
not required. The `click` parser still calls `consume_value` though
which calls our `prompt_for_value` and so hits the code that prints the
prompt info.

By moving that code after the check of whether the prompt function
should even be called, prevents the info to be displayed for these
options that don't actually need to be prompted for.