CLI: fix choices of `InteractiveOption` not being displayed (#5432)
The choices of interactive options were not being displayed in
interactive mode when the special character `?` was passed. This bug was
introduced when `click_completion` was dropped for the tab-completion
feature that ships with `click` itself as of `v8.0`. The complete method
changed from `complete` to `shell_complete` and the `get_help_message`
of the `InteractiveOption` was not properly updated.