`JobTemplate`: change `CodeInfo` to `JobTemplateCodeInfo` in `codes_info` (#5350)
When a `CalcJob` is run, the `presubmit` method will take the list of
`CodeInfo` instances from the `CalcInfo` returned by the method
`prepare_for_submission` and pass it to `Scheduler` implementation by
calling `get_submit_script`. The scheduler will then use the `CodeInfo`
instances to construct the run line.

It is not ideal that the `aiida.common.datastructures.CodeInfo` instance
is reused and passed to the scheduler for two reasons:

 * The `CalcJob.presubmit` alters the `CodeInfo` returned by the plugin
   implementation of `prepare_for_submission` and prepends MPI related
   parameters and the executable name. After these changes, the docs
   of the `CodeInfo` class are no longer respected.
 * By tightly coupling the data class that is used to communicate within
   the `CalcJob` class itself, as well as with the `Scheduler` plugin
   makes it difficult to add additional data that might be necessary
   for the `Scheduler` but would not be specified by the `CalcJob` plugin.

Both problems are solved by creating an independent and dedicated data
structure to communicate between `CalcJob` and `Scheduler. A dataclass
is added `aiida.schedulers.datastructures.JobTemplateCodeInfo`. The
`presubmit` method will now construct `JobTemplateCodeInfo` based on the
`CodeInfo` obtained from the `CalcInfo` and no longer modify them in
place. This will also allow to add new attributes to `JobTemplateCodeInfo`
in the future whenever needed.

This change could be breaking if there are `Scheduler` plugins out there
that rely on the type of the elements in the `JobTemplate.codes_info`
attribute, or when they try to access the `withmpi` or `code_uuid`
attributes since these are not present on `JobTemplateCodeInfo`. The
plugins that registered on the AiiDA plugin registry do not rely on
either of this, so we think the changes should have minimal impact on
existing code.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>