`CalcJob`: always call `Scheduler.parse_output`
When the functionality was added for a `Scheduler` plugin to parse the
output that was written to `stdout` and `stderr` or retrieved from a
specialized status command, it was decided to only call `parse_output`
if all three information streams were successfully retrieved. The
reasoning was especially that the `detailed_info` should be required as
that would return well-structured data and would allow to reliably
determine what had happened, whereas parsing free text from stdout and
stderr would be error prone.

Although probably a safe choice, the direct result was that setups where
the scheduler didn't have the necessary implementation to return the
`detailed_info`, no scheduler output parsing would be available. This
would lead to many OOM and OOW problems to go unnoticed and the engine
retrying to submit without any error handling.

Here we remove the requirement that all three information streams from
the parser, `detailed_info`, `stderr` and `stdout` should be known, but
that `parse_output` will always be called, with a default of `None` for
each. The advantage is that this will allow scheduler plugins to
implement parsing from `stderr`, on top of `detailed_info` wherever
applicable, to increase the chances of catching basic problems.

The major downside is that this is a backwards incompatible change for
scheduler plugins that rely on the assumption that the arguments always
have values that are not `None`.

This commit also fixes a bug in `CalcJob.parse_scheduler_output` where
it would pass the result of `node.get_option('scheduler_stderr')`
straight to `retrieved.get_object_content`. However, if the option was
not defined on the node, `get_option` will return `None` which will
result in a `TypeError` from the `get_object_content` call. This is
fixed by explicitly checking for `None` in which case a warning is
logged.