`SlurmScheduler`: use `stderr` in `parse_stdout` for OOW and OOM
The `CalcJob` implementation was changed to always call the
`parse_output` method of the scheduler, even if `detailed_info` is
`None`. This means that now we can attempt to parse errors from the
`stderr` as well.

Here we add simple regexes to try and detect OOM and OOW errors. They
return the exact same exit code as if they would have been detected from
the `detailed_info`. Note that since this is done with regexes, this
opens the door to false positives. It is not know how likely these are
to occur.