Dependencies: drop the `python-dateutil` library
This was only used in `aiida.common.timezone.isoformat_to_datetime` and
in the `aiida.restapi.common.utils` module to parse a datetime object
from a string in ISO format. Since Python 3.7 this functionality is
provided in the `datetime` standard library:

    https://docs.python.org/3/library/datetime.html#datetime.datetime.fromisoformat

The only relevant difference between the two implementations is that the
standard library is intentionally strict and will raise for invalid
formats whereas the `python-dateutil` will be lenient and try to guess
the intention in case of invalid formats. For both cases, we prefer not
to guess and potentially guess wrong, so the change is appropriate.