Simplify implementation of `aiida.common.hashing.get_random_string`
This function is used only to generate a password for the database user
when a new profile is setup through `verdi quicksetup`. Although this
password is not even critical, back in the early days when this was
added, the implementation for the method was taken from Django.

As of Python 3.6 though, there exists a much simpler solution that is
based on the `secrets` built in library. See the following for details:

    https://docs.python.org/3/library/secrets.html#recipes-and-best-practices