Fixing, simplifying and typing `aiida.common.timezone`
The `make_aware` method was actually bugged if the passed `timezone` was
not a `pytz.tzinfo` object but a standard `datetime.tzinfo`. The latter
does not have the `localize` method. The implementation checked for this
but outside the conditional still called the method, which would raise
an `AttributeError`.

Instead of fixing the small bug, the entire implementation is simplified
significantly. The implementation was borrowed from Django way in the
beginning of this project, but as of Python 3.6 `datetime.astimezone`
provides all the needed functionality as it will automatically retrieve
the timezone of the operating system when no timezone is specified. This
also means that we can now drop the `tzlocal` dependency as its only use
was determining the system's timezone. The `get_current_timezone` method
that wrapped it is dropped. The function now no longer excepts if a
datetime object is passed that is already aware, it will simply have the
timezone updated.

The `localtime` also was taken from Django originally and is also
simplified. It also drops the `timezone` argument because with it, it is
identical to `make_aware` rendering it obsolete. It should not allow
specifying a `timezone` but always use the timezone of the OS. Similar
to `make_aware`, this now also no longer excepts when an already aware
object is passed.

With these changes, the code no longer includes code taken from Django
and so its license can be dropped from `open_source_licenses.txt`.

Invocations of `timezone.delta` don't need to explicitly pass
`timezone.now` as second argument as that is already the default.

The use of `pytz` is reduced to a minimum and only used when there is no
equivalent functionality in the standard library. Instead of `pytz.utc`
use `datetime.timezone.utc`. For convenience this is forwarded in
`aiida.common.timezone`. The `pytz` library is now exclusively used to
convert a timezone name in string form into a `timezone` object. This
functionality is centralized in `aiida.common.timezone.timezone_from_name`.
This is also only used in `aiida.schedulers.datastructures` to convert
the serialized response from schedulers. If this were to be removed, the
`pytz` dependency could be dropped.

Finally, the `aiida.manage.configuration.settings.USE_TZ` settingm is
dropped. It was a relic of Django's configuration. It was hard-coded to
true, wasn't configurable and was only used in `aiida.common.timezone.now`
to make the returned object aware of the system's timezone if the setting
was set to True. This behavior is now hard-coded.