♻️  REFACTOR: Move link related methods to `Node.base.links`
This commit is part of the `Node` namespace restructure. It moves all
link related methods to a new `NodeLinks` class that is exposed through
the `Node.base.links` attribute.

A number of the `Node` subclasses for processes overrode the validation
methods `validate_incoming` and `validate_outgoing`. By changing the
design from straight-up inheritance to composition, and the link
validation functionality moving to the `NodeLinks` class, its is this
class that has to be overridden. To do so, the exact class to be used
for the `NodeLinks` class is made configurable on the `Node` class
through the `CLS_NODE_LINKS` class attribute. The subclasses can then
override the `NodeLinks` class and set this as the `CLS_NODE_LINKS`.

The `has_cached_links` property has been removed. It was only used
internally to check that when a process started running, all links had
been properly stored. This is a historic implementation detail, and
really what should be checked simply is that the node is stored. The
process should not start running before the node is stored and all links
are stored when the node is stored.