♻️  REFACTOR: privatize `validate_storability` and `verify_are_parents_stored`
These methods should only be used by the `Node` class itself and so are
best made private. Technically, `verify_are_parents_stored` is called
from one node instance on another and so should be public, but this just
one case and the methods are cluttering the namespace making it more
difficult to use for users in an interactive shell trying to discover
the method they are looking for.