CLI: catch `LockedProfileError` in `with_dbenv` decorator (#5487)
The `LockedProfileError` is thrown when a profile's storage is loaded
while it is locked by another process. When this is thrown during a CLI
command call, it is nice to convert it to a formatted message instead of
letting the exception bubble up and show the entire stacktrace.

The `aiida.cmdline.utils.decorators.with_dbenv` decorator is updated to
catch the `LockedProfileError` and call `echo.echo_critical`. This will
print the exception message and exit the command with a non-zero exit
status.

This fix does require that all CLI commands properly use `with_dbenv` if
they require the storage to be loaded. If they cannot use it for
whatever reason, the should ensure to catch the `LockedProfileError`
themselves and echo the exception message.