♻️  REFACTOR: rename `Entity.objects` to `Entity.collection` (#5490)
The `objects` class property would return a collection of the entities
and so the `collection` name would be more descriptive.

The `objects` property was abstract and implemented in each subclass of
`Entity` however the implementation was identical for each, save for the
type of `Collection` subclass to be used. Therefore, the property is
made concrete and implemented on `Collection` where it takes the
collection type to use from the `_CLS_COLLECTION` class attribute. This
is overridden in each `Entity` subclass to define the correct collection
implementation.