Tests: reinstate `test_daemon.py` in CI workflow (#5479)
This got removed in 5d2f15380417ce963544e209346b31eb37b0490d by accident
when moving the RPN workchains to the nightly build. The tests are
integration tests and take roughly 4 minutes so they are added to the
nightly build instead of the normal CI build as it would slow down PR
builds too much.