`Computer`: fallback on transport for `get_minimum_job_poll_interval` default (#5457)
The `get_minimum_job_poll_interval` returns a default when no value was
ever explicitly defined for the `Computer` instance. This is currently
provided by the `PROPERTY_MINIMUM_SCHEDULER_POLL_INTERVAL__DEFAULT`
attribute which is set to 10 seconds. This is a reasonable default for
most cases, but for the localhost it is probably excessive and can lead
to users waiting for a long time when running singular jobs on the
localhost.

The minimum job poll interval is not strictly a property of the
transport but in this case it is the transport that defines what happens
to be a sensible default. Therefore we still choose to add a new
attribute `DEFAULT_MINIMUM_JOB_POLL_INTERVAL` to `Transport` which the
`Computer` can use as a default if none is specified on the instance.
The `Transport` base class sets it to 10, which is backwards compatible
with the current behavior, and the `LocalTransport` then overrides it to
`0.1`. We choose a small finite number such that the waiting time in
most cases should not really be noticeable without running the risk that
the CPUs will really spin like crazy for jobs that run slightly longer.