Add option to use double quotes for `Code` and `Computer` CLI arguments
So far, the command line arguments of the code exeuction line in the
submit script written by the `Scheduler` plugin, would be wrapped by
single quotes to prevent special characters, such as spaces, from being
interpreted incorrectly. However, for certain use cases, the quoting has
an undesirable effect. For example, if an argument contains a bash
variable, indicated by a `$` sign, the single quotes will prevent it
from being dereferenced correctly.

A concrete use-case is the requested feature of being able to define a
Docker container as a `Code`, which needs certain arguments to be quoted
in double quotes. This PR adds a new property `use_double_quotes` to the
`Code` and `Computer` classes that determines whether arguments should
be quoted using single or double quotes.

The settings are communicated through the `cmdline_params` attribute of
a `JobTemplateCodeInfo` instance to the scheduler plugin. The plugin is
then responsible for respecting these booleans. To be able to
distinguish between the command line arguments of the computer and the
code, they are passed separately in `prepend_cmdline_params` and
`cmdline_params`, respectively.