`SandboxFolder`: decouple the location from the profile
The `SandboxFolder` by default would create the temporary directory in
the `sandbox` subfolder of the repository directory, defined by the
`repository_uri` key in the profile configuration.

However, this folder path should mainly serve to designate the location
of the file repository of the storage backend. So far the sandbox folder
hijacking the same key wasn't a problem as there was only one storage
backend implementation. However, when we allow other storage backend
implementations, having to keep this config key just for the sandbox
folder doesn't make sense. Additionally, the `SandboxFolder` can and
should be a generic utility that is not tightly coupled to the concept
of a `Profile`, as it was, by directly using it in the constructor.

The original reason for storing the sandbox folder in the folder of the
file repository is that this path is configurable and so users could
define a location on file system with plenty space available or a
filesystem that is in any other way more suited than the default
temporary directory of the operating system.

To keep the possibility for this configuration, the `SanboxFolder`
constructor now takes the optional `filepath` argument. If defined, the
sandbox folder will be created in this directory. By default, the folder
will be created in the default temporary directory of the operating
system. The `tempfile.mkdtemp` built in method is used for this purpose.