Fix `aiida.common.timezone.delta` having the wrong sign (#5493)
In the refactoring of 25ac2131c0eaa61baa8f722427ed5a01a6f55688, the
operands of the subtraction in `delta` were accidentally inversed. This
would lead to a negative delta and would manifest, for example, in all
relative time differences reported in the output of `verdi process list`
to be listed as `0s ago`.