DevOps: Change name of main branch from `develop/master` to `main`
With the release of `aiida-core==2.0`, it was decided to simplify the
branching model and only keep the `master` branch, dropping the
`develop` branch permanently.

In addition, the `master` branch is renamed to `main` following the name
change that was adopted by Github some time ago which now uses `main` as
the default branch name for new repositories.