Dependencies: update requirement `click~=8.1` (#5504)
A workaround was added in 186b16f419eca80022c084691c8a1cef407ab6ea
because the tab-completion in `click==8.0.x` would not autocomplete
option values starting with special characters. This behavior was fixed
in `click==8.1` so we can remove the workaround at the cost of upping
the lower version requirement.