Add the `InstalledCode` and `PortableCode` data plugins
Historically, the `Code` data plugin came in two flavors: codes that
represent an executable on an associated `Computer`, and codes that were
included in the repository of the node instance itself. For the latter,
when used in a `CalcJob` its files would be uploaded to the remote
working directory.

The use of a single class for these two different types of entities has
caused quite a bit of confusion, not only for end-users, but also for
developers as it leads to convoluted code.

Here, we add two new data plugins, `InstalledCode` and `PortableCode` that
are designed to implement the exact same behavior, but do so in a more
user-friendly and maintainable way by having them as separate classes.
The both subclass from `AbstractCode` which provides the functionality
that is shared between them.

For backwards compatibility, the original `Code` implementation is kept,
but it is moved to the `aiida.orm.nodes.data.code.legacy` module. The
idea is that this class is deprecated and users are encouraged to start
using the new classes. In a future version, the `Code` class will be
decommissioned and an automatic migration will be included to migrate
its existing entries to a `InstalledCode` or `PortableCode` instance.