Remove use of legacy `Code`
The `Code` plugin is used a lot and so cannot just be removed. Instead,
we keep it for now and have it subclass the new `AbstractCode`. This
way, the class already has the interface that it will have once the
deprecated interface is removed. This will allow users to start updating
their use of it in a backwards-compatible manner.

The codebase is updated to replace the use of the deprecated interface
with that of the new interface. No new instances of `Code` are created,
only instances of the new `RemoteCode` and `LocalCode`.

When the time comes to drop the legacy `Code` class, all that is needed
is a data migration that will update existing instances to either an
instance of `InstalledCode` or `PortableCode` depending on the attributes
that the instance defines.

The legacy methods and properties of `Code` plugin are deprecated.