CLI: Add the `DynamicEntryPointCommandGroup` command group
This custom implementation of `click.Group` will allow a CLI command to
dynamically generate its subcommands. The subcommands will be based on the
installed entry points for a specific entry point group. Optionally, the
entry points can be filtered using a regex pattern.

This approach was already in use to dynamically provide the commands to
configure computers with the various transport type plugins. The
functionality is generalized and added as a single class to the
`aiida.cmdline.groups.dynamic` module.

The code base already had one other implementation of `Group`, namely the
`VerdiCommandGroup`. For consistency, this is now also moved from
`aiida.cmdline.commands.cmd_verdi` to `aiida.cmdline.groups.verdi`.