CLI: Add the `verdi code create` command group
This is the successor of `verdi code setup`. The difference being that
where `verdi code setup` hardcoded the options and was designed to setup
both remote and local codes, `verdi code create` is a command group
that dynamically loads subcommands for each installed entry point that
is a subclass of the `AbstractCode` data plugin.

The advantage of the new approach is that it is much clearer for both
users as well as developers. The code and user interface for the code
setup command was getting quite complicated because a single command was
being used for two different purposes. By decoupling them, the interface
is much more intuitive and the code is much more easy to maintain.