Add the config option `storage.sandbox`
This config option can be used to define the directory into which all
sandbox folders will be created. This can be useful if the default
temporary directory of the operating system that would otherwise be
used is not suited, for example because it is on a slower file system or
it has insufficient empty space.

The option is passed to the `SandboxFolder` constructor whenever it is
used by AiiDA's internal code. Since the option requires the presence of
a configuration, the retrieval of the option is added as high up in the
API as possible. This is in order to not couple components to the
configuration that should remain independent.