fix: accept multiple identical entry points
When encountering multiple entrypoints of the same name in a given
group, still proceed if their values (where they are pointing) are
identical.

This makes it possible to e.g. install a version of AiiDA locally, over
a system-level installation of AiiDA that cannot be uninstalled (as long
as those versions do not change the value of an entry point).