Fix `get_entry_point` not raising even for duplicate entry points (#5531)
The `aiida.plugins.entry_point.get_entry_point` method is supposed to
raise `MultipleEntryPointError` if multiple entry points are registered
for the given group and name. However, since the recent change from
`reentry` to `importlib.metadata` the function was broken and when
multiple entry points were found, a random one would be returned. This
went unnoticed since the function was not being tested.

The direct problem is that the `names` property of the return value of
`eps().select()` eliminates duplicate names. So even for multiple entry
points, `len(eps().select().names)` would return `1` if all those entry
points had the same name, even though the `value` could be different.
The solution is to check for the length of the list itself that is
returned instead of checking the entry point names. 

There is one more subtlety that we account for. If multiple entry points
are found *but* they also all have the same value, i.e., they really
correspond to the same resource, we don't except but simply return one
of the entry points since they are all identical anyway.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>